20231002 별점 js ing

+++ kids-mall/kids-mall/src/main/java/team/lovelynephew/kidsmall/service/rating/RatingService.java
@@ -2,5 +2,5 @@ package team.lovelynephew.kidsmall.service.rating;
 
 public interface RatingService {
 
-	public int rating(int prdCode) throws Exception;
+	public double rating(int prdCode) throws Exception;
 }
